requirements.txt
openai==1.57.0
python-dotenv==1.0.1
pandas==2.2.2
scipy==1.14.1
openpyxl==3.1.5
jsonlines==4.0.0
pyyaml==6.0.2